+++ views/layouts/main.php
@@ -61,7 +61,7 @@ $bundle = ishizune\gentelella\assets\Asset::register($this);
                     <div id="sidebar-menu" class="main_menu_side hidden-print main_menu">
 
                         <div class="menu_section">
-                            <h3>General</h3>
+                            <h3>Menu</h3>
                             <?=
                                 \ishizune\gentelella\widgets\Menu::widget(
                                     [